[GUIS-39] Variety of UI changes made
This commit handles a variety of UI changes. These include:

- adding slider objects to modify the width, height, and gap length
  for the various pipe objects that the user interacts with
- fixing an issue where a given pipe could be scored multiple times
- increasing the default width of pipes as the levels progress

There are still unresolved issues with this build, but I want to
commit this changes, so I can begin working on development of the
shared game assets submodule. I will return to these UI issues once
that has been completed.